5-cdk-glue-job-trigger/requirements.txt
aws-cdk-lib
constructs
typeguard~=2.13.3